build: datasets.sh now with args

+++ datasets.sh
@@ -3,14 +3,27 @@
 set -e
 shopt -s extglob
 ROOT_DIR=$(pwd)
+TARGET=$1
 
-# EMNIST
-EMNIST_DIR="$ROOT_DIR/datasets/emnist"
-EMNIST_URL="https://biometrics.nist.gov/cs_links/EMNIST/gzip.zip"
-mkdir -p "$EMNIST_DIR"
-cd "$EMNIST_DIR"
-wget -q --show-progress "$EMNIST_URL" -O emnist.zip
-unzip -q -j emnist.zip
-rm emnist.zip
-rm !(emnist-digits-*|emnist-letters-*)
-gunzip -f *.gz
+# args validation
+if [ "$TARGET" != "all" ] && [ "$TARGET" != "emnist" ]; then
+    echo "Usage: ./datasets.sh [all|emnist]"
+    exit 1
+fi
+
+# download
+if [ "$TARGET" = "all" ] || [ "$TARGET" = "emnist" ]; then
+    EMNIST_DIR="$ROOT_DIR/datasets/emnist"
+    EMNIST_URL="https://biometrics.nist.gov/cs_links/EMNIST/gzip.zip"
+    rm -rf "$EMNIST_DIR"
+    mkdir -p "$EMNIST_DIR"
+    cd "$EMNIST_DIR"
+    wget -q --show-progress "$EMNIST_URL" -O emnist.zip
+    unzip -q -j emnist.zip
+    rm emnist.zip
+    rm !(emnist-digits-*|emnist-letters-*)
+    gunzip -f *.gz
+fi
+# if [ "$TARGET" = "all" ] || [ "$TARGET" = "" ]; then
+#     # TODO
+# fi
